Deprecates this crate in favor of proxy-wasm
This crate both depends on a personal library and is also unmaintained.

Envoy users should be pointed to a maintained repository instead.
Currently
[proxy-wasm](https://github.com/proxy-wasm/proxy-wasm-rust-sdk) is
maintained with recent releases by @PiotrSikora. While this includes
some functionality not present there, maintained and without personal
dependencies is a better path for users than otherwise.

Signed-off-by: Adrian Cole <adrian@tetrate.io>